+++ src/Event/ModelCreated.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Pulsar\Event;
+
+final class ModelCreated extends AbstractEvent
+{
+    const NAME = 'model.created';
+}