+++ src/Event/ModelDeleted.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Pulsar\Event;
+
+final class ModelDeleted extends AbstractEvent
+{
+    const NAME = 'model.deleted';
+}