+++ src/Event/ModelDeleting.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Pulsar\Event;
+
+final class ModelDeleting extends AbstractEvent
+{
+    const NAME = 'model.deleting';
+}