+++ src/Event/ModelUpdated.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Pulsar\Event;
+
+final class ModelUpdated extends AbstractEvent
+{
+    const NAME = 'model.updated';
+}