+++ src/Event/ModelUpdating.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Pulsar\Event;
+
+final class ModelUpdating extends AbstractEvent
+{
+    const NAME = 'model.updating';
+}